issue-349 types refactoring
