+++ scroller/public-api.ts
@@ -1,4 +1,5 @@
 export { UiScrollModule } from './src/ui-scroll.module';
 export { UiScrollDirective } from './src/ui-scroll.directive';
-export { IAdapter, IDatasource, Datasource } from './src/ui-scroll.datasource';
+export { Datasource } from './src/ui-scroll.datasource';
+export { IAdapter, IDatasource } from './src/types';
 export { Routines, SizeStrategy } from './src/vscroll';
