+++ scroller/src/ui-scroll.datasource.ts
@@ -4,42 +4,11 @@ import {
   makeDatasource,
   AdapterPropName,
   EMPTY_ITEM,
-  IDatasource,
-  IDatasourceConstructed,
   IReactivePropConfig,
   IAdapterConfig,
-  IAdapterItem,
-  IAdapter
+  IAdapterItem
 } from './vscroll';
-
-interface IReactiveOverride<Item = unknown> {
-  init$: Subject<boolean>;
-  isLoading$: Subject<boolean>;
-  loopPending$: Subject<boolean>;
-  firstVisible$: BehaviorSubject<IAdapterItem<Item>>;
-  lastVisible$: BehaviorSubject<IAdapterItem<Item>>;
-  bof$: Subject<boolean>;
-  eof$: Subject<boolean>;
-}
-
-type _Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
-
-interface IAngularAdapter<Data = unknown>
-  extends _Omit<IAdapter<Data>, keyof IReactiveOverride<Data>>,
-    IReactiveOverride<Data> {}
-
-interface IAngularDatasourceParams<Data = unknown>
-  extends _Omit<IDatasource<Data>, 'adapter'> {}
-
-interface IAngularDatasource<Data = unknown>
-  extends _Omit<IDatasource<Data>, 'adapter'> {
-  adapter?: IAngularAdapter<Data>;
-}
-
-interface IAngularDatasourceConstructed<Data = unknown>
-  extends _Omit<IDatasourceConstructed<Data>, 'adapter'> {
-  adapter: IAngularAdapter<Data>;
-}
+import { AngularDatasourceClass } from './types';
 
 const getBooleanSubjectPropConfig = (): IReactivePropConfig => ({
   source: new Subject<boolean>(),
@@ -65,16 +34,8 @@ const getAdapterConfig = (): IAdapterConfig => ({
   }
 });
 
-type AngularDatasourceClass<> = new <Data>(
-  params: IAngularDatasourceParams<Data>
-) => IAngularDatasourceConstructed<Data>;
-
 const AngularDatasource = makeDatasource(
   getAdapterConfig
 ) as AngularDatasourceClass;
 
-export {
-  IAngularAdapter as IAdapter,
-  IAngularDatasource as IDatasource,
-  AngularDatasource as Datasource
-};
+export { AngularDatasource as Datasource };
